+++ lib/ui/directory/DirectoryWidgets.dart
@@ -262,17 +262,18 @@ class _DirectoryAccountListCardState extends State<DirectoryAccountListCard> {
 
 // DirectoryAccountBusinessCard
 
-class DirectoryAccountBusinessCard extends StatefulWidget {
+class DirectoryAccountContactCard extends StatefulWidget {
   final Auth2PublicAccount? account;
   final String? accountId;
+  final bool printMode;
 
-  DirectoryAccountBusinessCard({super.key, this.account, this.accountId });
+  DirectoryAccountContactCard({super.key, this.account, this.accountId, this.printMode = false });
 
   @override
-  State<StatefulWidget> createState() => _DirectoryAccountBusinessCardState();
+  State<StatefulWidget> createState() => _DirectoryAccountContactCardState();
 }
 
-class _DirectoryAccountBusinessCardState extends State<DirectoryAccountBusinessCard> {
+class _DirectoryAccountContactCardState extends State<DirectoryAccountContactCard> {
 
   Auth2PublicAccount? _account;
   bool _loadingAccount = false;
@@ -280,9 +281,10 @@ class _DirectoryAccountBusinessCardState extends State<DirectoryAccountBusinessC
   Auth2UserProfile? get _profile => _account?.profile;
 
   String? get _photoImageUrl => StringUtils.isNotEmpty(_profile?.photoUrl) ?
-    Content().getUserPhotoUrl(accountId: _account?.id, type: UserProfileImageType.medium) : null;
+    Content().getUserPhotoUrl(accountId: _account?.id, type: _photoImageType) : null;
 
   double get _photoImageSize => MediaQuery.of(context).size.width / 3;
+  UserProfileImageType get _photoImageType => widget.printMode ? UserProfileImageType.defaultType : UserProfileImageType.medium;
 
   Map<String, String>? get _photoAuthHeaders => DirectoryProfilePhotoUtils.authHeaders;
 
@@ -366,7 +368,7 @@ class _DirectoryAccountBusinessCardState extends State<DirectoryAccountBusinessC
         Expanded(child:
           Center(child:
             Row(mainAxisSize: MainAxisSize.min, crossAxisAlignment: CrossAxisAlignment.start, children: [
-              if (_profile?.pronunciationUrl?.isNotEmpty == true)
+              if (!widget.printMode && (_profile?.pronunciationUrl?.isNotEmpty == true))
                 DirectoryPronunciationButton.spacer(),
               Column(mainAxisSize: MainAxisSize.min, children: [
                 Padding(padding: EdgeInsets.only(top: 16), child:
@@ -375,7 +377,7 @@ class _DirectoryAccountBusinessCardState extends State<DirectoryAccountBusinessC
                 if (_profile?.pronouns?.isNotEmpty == true)
                   Text(_profile?.pronouns ?? '', style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
               ]),
-              if (_profile?.pronunciationUrl?.isNotEmpty == true)
+              if (!widget.printMode && (_profile?.pronunciationUrl?.isNotEmpty == true))
                 DirectoryPronunciationButton(url: _profile?.pronunciationUrl),
             ],),
           ),
@@ -392,7 +394,7 @@ class _DirectoryAccountBusinessCardState extends State<DirectoryAccountBusinessC
         if (_profile?.pronouns?.isNotEmpty == true)
           Text(_profile?.pronouns ?? '', style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
       ]),
-      if (_profile?.pronunciationUrl?.isNotEmpty == true)
+      if (!widget.printMode && (_profile?.pronunciationUrl?.isNotEmpty == true))
         DirectoryPronunciationButton(url: _profile?.pronunciationUrl),
     ],),
   ]);
@@ -445,7 +447,7 @@ class DirectoryAccountPopupCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => Stack(children: [
-    DirectoryAccountBusinessCard(account: account, accountId: accountId,),
+    DirectoryAccountContactCard(account: account, accountId: accountId,),
     Positioned.fill(child:
       Align(alignment: Alignment.topRight, child:
         InkWell(onTap: () => _onTapClose(context), child: