+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Edit a message [#4579](https://github.com/rokwire/illinois-app/issues/4572).
 - Create recurring events - task in progress [#4599](https://github.com/rokwire/illinois-app/issues/4599).
+- Implemented Save to Photos and Share Virtual Card commands for Profile [#4605](https://github.com/rokwire/illinois-app/issues/4605).
 ### Changed
 - Updated Connections and Directory section descriptions [#4603](https://github.com/rokwire/illinois-app/issues/4603).